+++ src/services/SegmentsService.php
@@ -10,13 +10,10 @@ use Craft;
 use craft\base\Component;
 use craft\helpers\Db;
 use craft\records\Element_SiteSettings;
-use craft\web\View;
 use putyourlightson\campaign\elements\ContactElement;
 use putyourlightson\campaign\elements\db\ContactElementQuery;
 use putyourlightson\campaign\elements\SegmentElement;
-use RuntimeException;
 use Twig\Error\LoaderError;
-use Twig\Error\RuntimeError;
 use Twig\Error\SyntaxError;
 
 /**